backend/requirements.txt
django==3.2.23
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pymysql==1.1.0
django-cors-headers==4.3.0
django-axes==5.39.0
Pillow==10.2.0
gunicorn==21.2.0
django-ckeditor==6.7.0
django-storages==1.14.3
